requirements.txt
Flask==3.0.3
azure-storage-blob==12.19.1
msal==1.27.0
gunicorn
python-dotenv
requests==2.31.0
